requirements.txt
Flask>=2.0,<3.0
flask-cors==3.0.10
pymongo==3.12.1
marshmallow==3.22.0
requests==2.26.0
python-dotenv==0.19.2
